# target/classes/openapi.yaml
openapi: 3.0.0
info:
  title: Repair Tips API
  description: API Backend para Sistema de Consulta de Dicas de Conserto de Equipamentos
  version: 1.0.0
  contact:
    name: API Support
    url: https://example.com
  license:
    name: MIT
servers:
  - url: http://localhost:8080/api
    description: Development Server
  - url: https://api.example.com/api
    description: Production Server
...
paths:
  /fabricantes:
    get:
      tags:
        - Fabricantes
      summary: Listar todos os fabricantes
      operationId: getFabricantes
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Lista de fabricantes retornada com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageFabricanteResponseDTO'
    post:
      tags:
        - Fabricantes
      summary: Criar novo fabricante
      operationId: createFabricante
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FabricanteRequestDTO'
      responses:
        '201':
          description: Fabricante criado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FabricanteResponseDTO'
        '400':
          description: Dados inválidos
  /fabricantes/{id}:
    get:
      tags:
        - Fabricantes
      summary: Obter fabricante por ID
      operationId: getFabricanteById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Fabricante encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FabricanteResponseDTO'
        '404':
          description: Fabricante não encontrado
    put:
      tags:
        - Fabricantes
      summary: Atualizar fabricante
      operationId: updateFabricante
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FabricanteRequestDTO'
      responses:
        '200':
          description: Fabricante atualizado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FabricanteResponseDTO'
        '404':
          description: Fabricante não encontrado
    delete:
      tags:
        - Fabricantes
      summary: Deletar fabricante
      operationId: deleteFabricante
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Fabricante deletado com sucesso
        '404':
          description: Fabricante não encontrado
  /fabricantes/nome/{nome}:
    get:
      tags:
        - Fabricantes
      summary: Buscar fabricante por nome
      operationId: getFabricanteByNome
      parameters:
        - name: nome
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Fabricante encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FabricanteResponseDTO'
        '404':
          description: Fabricante não encontrado
...
  /pecas:
    get:
      tags:
        - Peças
      summary: Listar todas as peças
      operationId: getPecas
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Lista de peças
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagePecaResponseDTO'
    post:
      tags:
        - Peças
      summary: Criar nova peça
      operationId: createPeca
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PecaRequestDTO'
      responses:
        '201':
          description: Peça criada com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PecaResponseDTO'
  /pecas/{id}:
    get:
      tags:
        - Peças
      summary: Obter peça por ID
      operationId: getPecaById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Peça encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PecaResponseDTO'
    put:
      tags:
        - Peças
      summary: Atualizar peça
      operationId: updatePeca
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PecaRequestDTO'
      responses:
        '200':
          description: Peça atualizada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PecaResponseDTO'
    delete:
      tags:
        - Peças
      summary: Deletar peça
      operationId: deletePeca
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Peça deletada com sucesso
components:
  schemas:
...
    FabricanteRequestDTO:
      type: object
      required:
        - nome
        - sigla
      properties:
        nome:
          type: string
          minLength: 3
          maxLength: 100
          description: Nome do fabricante
        sigla:
          type: string
          minLength: 2
          maxLength: 5
          description: Sigla do fabricante
    FabricanteResponseDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: ID do fabricante
        nome:
          type: string
          description: Nome do fabricante
        sigla:
          type: string
          description: Sigla do fabricante
...
    PecaRequestDTO:
      type: object
      required:
        - nome
        - partNumber
        - categoria
        - equipamentoId
      properties:
        nome:
          type: string
          description: Nome da peça
        partNumber:
          type: string
          description: Número da parte
        categoria:
          type: string
          description: Categoria da peça
        equipamentoId:
          type: integer
          format: int64
          description: ID do equipamento
    PecaResponseDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nome:
          type: string
        partNumber:
          type: string
        categoria:
          type: string
        equipamentoId:
          type: integer
          format: int64
    PageFabricanteResponseDTO:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/FabricanteResponseDTO'
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
        currentPage:
          type: integer
...
    PagePecaResponseDTO:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/PecaResponseDTO'
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
        currentPage:
          type: integer